requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
scipy
scikit-learn
nltk
lexicalrichness
textstat
language-tool-python
aiofiles
nemo_toolkit[asr]